Moving Literal inspections to their own file
