+++ mulang.cabal
@@ -47,6 +47,7 @@ library
     Language.Mulang.Inspector.Generic
     Language.Mulang.Inspector.Generic.Duplication
     Language.Mulang.Inspector.Generic.Smell
+    Language.Mulang.Inspector.Literal
     Language.Mulang.Inspector.Logic
     Language.Mulang.Inspector.ObjectOriented
     Language.Mulang.Inspector.ObjectOriented.Polymorphism
