+++ src/Language/Mulang/Inspector.hs
@@ -1,22 +1,24 @@
 module Language.Mulang.Inspector (
   Inspection,
   BoundInspection,
-  module Language.Mulang.Inspector.Contextualized,
   module Language.Mulang.Inspector.Combiner,
+  module Language.Mulang.Inspector.Contextualized,
+  module Language.Mulang.Inspector.Functional,
   module Language.Mulang.Inspector.Generic,
-  module Language.Mulang.Inspector.Query,
-  module Language.Mulang.Inspector.Typed,
+  module Language.Mulang.Inspector.Literal,
+  module Language.Mulang.Inspector.Logic,
   module Language.Mulang.Inspector.ObjectOriented,
   module Language.Mulang.Inspector.ObjectOriented.Polymorphism,
-  module Language.Mulang.Inspector.Functional,
-  module Language.Mulang.Inspector.Logic,
-  module Language.Mulang.Inspector.Procedural) where
+  module Language.Mulang.Inspector.Procedural,
+  module Language.Mulang.Inspector.Query,
+  module Language.Mulang.Inspector.Typed) where
 
-import Language.Mulang.Inspector.Combiner
 import Language.Mulang.Inspector.Bound (BoundInspection)
+import Language.Mulang.Inspector.Combiner
 import Language.Mulang.Inspector.Contextualized
 import Language.Mulang.Inspector.Functional
 import Language.Mulang.Inspector.Generic
+import Language.Mulang.Inspector.Literal
 import Language.Mulang.Inspector.Logic
 import Language.Mulang.Inspector.ObjectOriented
 import Language.Mulang.Inspector.ObjectOriented.Polymorphism
