+++ src/Language/Mulang/Inspector/Generic.hs
@@ -22,9 +22,6 @@ module Language.Mulang.Inspector.Generic (
   usesExceptions,
   usesFor,
   usesIf,
-  usesLiteral,
-  isLiteral,
-  areLiterals,
   usesYield) where
 
 import Language.Mulang.Ast
@@ -38,31 +35,9 @@ import Language.Mulang.Inspector.Query (inspect, select)
 import Data.Maybe (listToMaybe)
 import Data.List.Extra (has)
 
-import Text.Read (readMaybe)
-
 -- | Inspection that tells whether an expression is equal to a given piece of code after being parsed
 parses :: (Code -> Expression) -> Code -> Inspection
 parses parser code = (== (parser code))
-
-type MultiInspection = [Expression] -> Bool
-
-anyExpressions :: MultiInspection
-anyExpressions = const True
-
-usesLiteral :: Code -> Inspection
-usesLiteral value = containsExpression (isLiteral value)
-
-isLiteral :: Code -> Inspection
-isLiteral value = f
-  where f MuNil              = value == "Nil"
-        f (MuNumber number)  = (readMaybe value) == Just number
-        f (MuBool bool)      = (readMaybe value) == Just bool
-        f (MuString string)  = (readMaybe value) == Just string
-        f (MuChar char)      = (readMaybe value) == Just char
-        f (MuSymbol string)  = (readMaybe value) == Just ("#" ++ string)
-
-areLiterals :: [Code] -> MultiInspection
-areLiterals codes expressions = and (zipWith isLiteral codes expressions)
 
 assigns :: BoundInspection
 assigns = assignsWith anyExpression
