+++ src/Language/Mulang/Inspector/Primitive.hs
@@ -1,10 +1,12 @@
 module Language.Mulang.Inspector.Primitive (
   anyExpression,
+  anyExpressions,
   containsExpression,
   containsDeclaration,
   containsBody,
   matchesType,
-  Inspection) where
+  Inspection,
+  MultiInspection) where
 
 import           Language.Mulang.Ast
 import           Language.Mulang.Identifier (IdentifierPredicate)
@@ -13,9 +15,13 @@ import           Language.Mulang.Generator (expressions, equationBodies, declara
 import           Data.List.Extra (has)
 
 type Inspection = Expression -> Bool
+type MultiInspection = [Expression] -> Bool
 
 anyExpression :: Inspection
 anyExpression = const True
+
+anyExpressions :: MultiInspection
+anyExpressions = const True
 
 containsExpression :: Inspection -> Inspection
 containsExpression f = has f expressions
